self invocation 문제로 비즈니스 로직을 외부 클래스로 분리
