+++ src/main/java/flab/gotable/service/ReservationService.java
@@ -1,6 +1,5 @@
 package flab.gotable.service;
 
-import flab.gotable.domain.entity.Reservation;
 import flab.gotable.dto.request.ReservationRequestDto;
 import flab.gotable.dto.response.ReservationResponseDto;
 import flab.gotable.exception.*;
@@ -11,25 +10,22 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class ReservationService {
+
+    private final ReservationBusinessService reservationBusinessService;
     private final ReservationMapper reservationMapper;
     private final MemberMapper memberMapper;
     private final StoreMapper storeMapper;
 
     @Transactional(isolation = Isolation.READ_COMMITTED)
     public ReservationResponseDto reservePessimisticLock(ReservationRequestDto reservationRequestDto) {
-        final long restaurantId = reservationRequestDto.getRestaurantId();
-        final LocalDateTime reservationStartTime = reservationRequestDto.getReservationStartTime();
-        final LocalDateTime reservationEndTime = reservationRequestDto.getReservationEndTime();
 
         validateReservationRequest(reservationRequestDto);
 
@@ -38,24 +34,13 @@ public class ReservationService {
             // 락 획득 실패
             throw new LockFailureException(ErrorCode.LOCK_ACQUISITION_FAILED, ErrorCode.LOCK_ACQUISITION_FAILED.getMessage());
         }
-        // 예약하려는 시간에 다른 사용자가 예약한 경우
-        if (!isReservationAvailable(restaurantId, reservationStartTime, reservationEndTime)) {
-            throw new DuplicatedReservationException(ErrorCode.DUPLICATED_RESERVATION_TIME, ErrorCode.DUPLICATED_RESERVATION_TIME.getMessage());
-        }
 
-        // 예약하려는 시간이 일반 또는 영업 스케줄 중에 존재하지 않는 경우
-        if (!isExistSchedule(restaurantId, reservationStartTime, reservationEndTime)) {
-            throw new ScheduleNotFoundException(ErrorCode.RESERVATION_TIME_NOT_FOUND, ErrorCode.RESERVATION_TIME_NOT_FOUND.getMessage());
-        }
-
-        Reservation reservation = ReservationRequestDto.toEntity(reservationRequestDto);
-        reservationMapper.saveReservation(reservation);
-
-        return new ReservationResponseDto(reservation);
+        return reservationBusinessService.executePessimisticLockReservation(reservationRequestDto);
     }
 
     @Transactional
     public ReservationResponseDto reserveNamedLock(ReservationRequestDto reservationRequestDto) {
+
         validateReservationRequest(reservationRequestDto);
 
         String lockName = "reservation_" + reservationRequestDto.getRestaurantId();
@@ -72,7 +57,7 @@ public class ReservationService {
         }
 
         try {
-            return executeReservation(reservationRequestDto);
+            return reservationBusinessService.executeNamedLockReservation(reservationRequestDto);
         } finally {
             Integer releaseResult = reservationMapper.releaseNamedLock(lockName);
 
@@ -81,28 +66,6 @@ public class ReservationService {
                 log.warn("Failed to release lock: {}", lockName);
             }
         }
-    }
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public ReservationResponseDto executeReservation(ReservationRequestDto reservationRequestDto) {
-        final long restaurantId = reservationRequestDto.getRestaurantId();
-        final LocalDateTime reservationStartTime = reservationRequestDto.getReservationStartTime();
-        final LocalDateTime reservationEndTime = reservationRequestDto.getReservationEndTime();
-
-        // 예약하려는 시간에 다른 사용자가 예약한 경우
-        if (!isReservationAvailable(restaurantId, reservationStartTime, reservationEndTime)) {
-            throw new DuplicatedReservationException(ErrorCode.DUPLICATED_RESERVATION_TIME, ErrorCode.DUPLICATED_RESERVATION_TIME.getMessage());
-        }
-
-        // 예약하려는 시간이 일반 또는 영업 스케줄 중에 존재하지 않는 경우
-        if (!isExistSchedule(restaurantId, reservationStartTime, reservationEndTime)) {
-            throw new ScheduleNotFoundException(ErrorCode.RESERVATION_TIME_NOT_FOUND, ErrorCode.RESERVATION_TIME_NOT_FOUND.getMessage());
-        }
-
-        Reservation reservation = ReservationRequestDto.toEntity(reservationRequestDto);
-        reservationMapper.saveReservation(reservation);
-
-        return new ReservationResponseDto(reservation);
     }
 
     private void validateReservationRequest(ReservationRequestDto reservationRequestDto) {
@@ -154,20 +117,5 @@ public class ReservationService {
         if(startTime.isBefore(LocalDateTime.now())) {
             throw new InvalidReservationTimeException(ErrorCode.PAST_RESERVATION_TIME, ErrorCode.PAST_RESERVATION_TIME.getMessage());
         }
-    }
-
-    private boolean isReservationAvailable(long restaurantId, LocalDateTime reservationStartTime, LocalDateTime reservationEndTime) {
-        return !reservationMapper.isDuplicatedReservation(restaurantId, reservationStartTime, reservationEndTime);
-    }
-
-    private boolean isExistSchedule(long restaurantId, LocalDateTime reservationStartTime, LocalDateTime reservationEndTime) {
-        final LocalTime startTime = reservationStartTime.toLocalTime();
-        final LocalTime endTime = reservationEndTime.toLocalTime();
-        final String dayOfWeek = reservationStartTime.getDayOfWeek().toString();
-
-        boolean isExistDailySchedule = reservationMapper.isExistDailySchedule(restaurantId, dayOfWeek, startTime, endTime);
-        boolean isExistSpecificSchedule = reservationMapper.isExistSpecificSchedule(restaurantId, reservationStartTime.toLocalDate(), startTime, endTime);
-
-        return isExistDailySchedule || isExistSpecificSchedule;
     }
 }
